+++ MoviewRecommend.DAL/Entities/Movie.cs
@@ -8,6 +8,6 @@ namespace MoviewRecommend.DAL.Entities
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int AverageScore { get; set; }
+        public decimal AverageScore { get; set; }
     }
 }
